Support non-scalar non-string collection keys
Tokensearch.js appears to have assumed that each of the collection keys
always represented text (which is an ideal situation). I needed to
support a case where maybe the collections keys were an array or an
object.

This is by no means a "complete" solution, and it will continue to be
iterated on.